requirements.txt
Flask
opencv-python  
numpy
pandas
os
Flask-Bootstrap
random
time
google-auth  
telebot
matplotlib 
tensorflow 
keras 
pydot